Split model to types and actions.
